+++ src/defaults.js
@@ -1,0 +1,9 @@
+(() => {
+  "use strict";
+  window.__WORKTIME_AUTOFILL_DEFAULTS__ = Object.freeze({
+    standardStart: "08:45",
+    standardEnd: "17:15",
+    standardBreak: "00:45",
+    autoOnLoad: false,
+  });
+})();